feat: implement Transaction class and debt minimization
The Transaction class was added to the model. A transaction contains who paid, the amount paid, the currency, a category, a date and time, and it specifies the way the debt is shared among the debtors through a list of DebtorShares. Each DebtorShare models how much a person owes.
The Group class was updated to include a collection of Debts between the members of the groups.
Whenever a transaction is added or removed, the debts are minimized through the "minimize_debts" service function.

Unit tests using pytest.